+++ src/Bridge/Doctrine/DoctrineBridge.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MakinaCorpus\QueryBuilder\Bridge\Doctrine;
+
+class DoctrineBridge extends DoctrineQueryBuilder
+{
+}